configmgr: ibm-clang64 + Linux/WSL builds + cross-platform tests
Modernize the `configmgr` build for IBM Open XL C/C++ (`ibm-clang64`) on
z/OS 3.1+ while keeping the `xlclang` path unchanged, and add a Linux/WSL
build mode (useful for catching cross-platform issues before they hit z/OS).
Both z/OS compilers and the Linux build produce a working `configmgr` with
zero warnings or errors.

Branch also lands a portable POSIX-shell regression suite under
`tests/configmgr/cli/` (128 assertions, 6 suites) and extends the existing
`tests/configmgr/extract/` and `tests/commenttest/` exercises to run under
both z/OS compilers AND Linux/WSL with byte-identical output verified
across all three.

Code changes
------------
- build/build_cmgr_clang.sh -- new build script (zos | zos-cross | linux modes).
- c/embeddedjs.c -- fixed-point iteration so YAML \${{ }} forward refs resolve correctly.
- c/configmgr.c -- remove leftover shoeSize:11 debug field from validate().
- c/dynalloc.c -- quiet -Wpointer-to-int-cast on SVC 99 plist sites via (uintptr_t)
  interposition; fix memset(plist, 0, sizeof(plist)) to sizeof(*plist) (real bug;
  SVC 99 was being invoked with stale tail bytes).
- c/psxskt.c -- warning cleanup; bug fix in getLocalHostName (printf had 3
  conversions, 2 args); replace mistyped setSocketReuseAddr(&sd, ...) with direct
  setsockopt() on the raw fd.
- c/signalcontrol.c -- fix BPX1SIA/BPX4SIA arg types (int64 oldFlags -> int;
  (char *)&zero for user_data); surfaced by zss's xlclang build pulling this TU
  in with the new strict prototypes.
- c/zosaccounts.c -- include strict BPX prototypes + <stdlib.h>; cast string args
  ((char *)userName/password/etc.) and BPX-returned address ((char *)(uintptr_t)
  returnValue) to silence type-mismatch warnings; fix real bug in userIdGet
  (atoi(userId) -> atoi(string); was returning -1 for all-digit inputs).
- h/signalcontrol.h -- #ifndef guards on SA_*/SIG_* macros so they coexist with
  <signal.h>.
- h/zowetypes.h -- guard __ZOWE_OS_LINUX define with #ifndef.

New / extended tests
--------------------
- tests/configmgr/cli/ -- 6 suites, 128 assertions, green on all three toolchains.
  known_issues/ runner for platform-conditional pins.
- tests/configmgr/extract/index.sh -- guard relaxed; runs cross-platform.
- tests/commenttest/build.sh -- now xlclang|clang|linux|both modes; three
  binaries with byte-identical output.
- tests/clang_readiness/ -- signalcontrol.c added to COMMON_C (compile-side
  regression guard); new test_signalcontrol exercises install path at runtime
  on both compilers.

+++ c/http.c
@@ -196,7 +196,7 @@ int writeFully(Socket *socket, char *buffer, int len){
 
   /* this is a sanity check for a bug that once happened on ZOS */
 #ifdef __ZOWE_OS_ZOS
-  if (((int)socket) < 0x1000000){
+  if ((uint64)socket < 0x1000000){
     return 0;
   }
 #endif
